requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.23.0
scikit-learn>=1.1.0
matplotlib>=3.6.0
PyYAML>=6.0
